task check

+++ 15-binary_tree_is_full.c
@@ -1,3 +1,5 @@
+#include "binary_trees.h"
+
 /**
  * binary_tree_is_full - checks if a binary tree is full
  * @tree: pointer to the node
@@ -6,7 +8,7 @@
 
 int binary_tree_is_full(const binary_tree_t *tree)
 {
-	size_t first, second;
+	int first, second;
 
 	if (tree == NULL)
 		return (0);
